Avoid using class_eval and just include a module
