+++ activejob/lib/active_job.rb
@@ -37,4 +37,5 @@ module ActiveJob
   autoload :ConfiguredJob
   autoload :TestCase
   autoload :TestHelper
+  autoload :QueryTags
 end
