+++ activejob/lib/active_job/query_tags.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module ActiveJob
+  module QueryTags # :nodoc:
+    extend ActiveSupport::Concern
+
+    included do
+      around_perform :expose_job_to_query_logs
+    end
+
+    private
+      def expose_job_to_query_logs(&block)
+        ActiveRecord::QueryLogs.set_context(job: self, &block)
+      end
+  end
+end
